[ADD] disbursement_assignment_kmitl: extract assignment feature from disbursement
Move the officer-assignment and return-for-correction feature out of the
disbursement core into a new standalone addon, reversing the earlier merge
(#959) that had absorbed the former disbursement_assignment_kmitl module.

New module disbursement_assignment_kmitl (depends: disbursement):
- models: assignment_rule + disbursement.request extension (auto-assign,
  assign/unassign, return-for-correction, workflow overrides)
- wizards: assign officer, return request
- views: request form/tree/search inherit, assignment rules, verification menus
- security: group_disbursement_verification + assignment ACL rows
- tests + i18n/th.po

disbursement:
- drop the assignment models, wizards, views, verification group and ACL rows
- move the assignment translations out of i18n/th.po
- bump version to 16.0.5.0.0

agx_approval_disbursement (only downstream consumer of the feature):
- depend on disbursement_assignment_kmitl
- retarget the inherited view ref to the new module

No data migration: this is a code-level split for fresh installs; databases
that already ran #959 should be reinstalled.
