+++ disbursement/__manifest__.py
@@ -2,7 +2,7 @@
 
 {
     "name": "Disbursement",
-    "version": "16.0.4.2.0",
+    "version": "16.0.5.0.0",
     "category": "Disbursement",
     "license": "LGPL-3",
     "author": "KMITL",
@@ -22,18 +22,13 @@
         "security/ir.model.access.csv",
         "data/sequence.xml",
         "wizards/disbursement_exception_confirm.xml",
-        "wizards/assign_officer_wizard_views.xml",
-        "wizards/return_request_wizard_views.xml",
         "views/disbursement_request_views.xml",
-        "views/disbursement_request_assignment_views.xml",
         "views/budget_commitment_views.xml",
         "views/exception_rule_views.xml",
-        "views/assignment_rule_views.xml",
         "report/paperformat.xml",
         "report/report_disbursement_request_action.xml",
         "report/report_disbursement_request.xml",
         "views/portal_templates.xml",
-        "views/disbursement_verification_menus.xml",
     ],
     "assets": {
         "web.assets_frontend": [
