+++ disbursement/i18n/th.po
@@ -989,12 +989,6 @@ msgid "Net Total"
 msgstr "มูลค่าสุทธิ"
 
 #. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__active
-msgid "Active"
-msgstr "ใช้งาน"
-
-#. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__activity_analytic_id
 #: model:ir.model.fields,field_description:disbursement.field_disbursement_request__activity_analytic_id
 msgid "Activity"
 msgstr "กิจกรรม"
@@ -1026,7 +1020,6 @@ msgid "Bill Posted"
 msgstr "ตั้งหนี้แล้ว"
 
 #. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__department_analytic_id
 #: model:ir.model.fields,field_description:disbursement.field_disbursement_request__department_analytic_id
 msgid "Department"
 msgstr "ส่วนงาน"
@@ -1047,7 +1040,6 @@ msgid "Done"
 msgstr "เสร็จสิ้น"
 
 #. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__fund_analytic_id
 #: model:ir.model.fields,field_description:disbursement.field_disbursement_request__fund_analytic_id
 msgid "Fund"
 msgstr "กองทุน"
@@ -1079,7 +1071,6 @@ msgid "Partner Information"
 msgstr "ข้อมูลคู่ค้า"
 
 #. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__partner_type_id
 #: model:ir.model.fields,field_description:disbursement.field_disbursement_request__partner_type
 msgid "Partner Type"
 msgstr "จ่ายตาม"
@@ -1149,7 +1140,6 @@ msgid "Signed"
 msgstr "ลงนามแล้ว"
 
 #. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__source_analytic_id
 #: model:ir.model.fields,field_description:disbursement.field_disbursement_request__source_analytic_id
 msgid "Source"
 msgstr "แหล่งเงิน"
@@ -1291,157 +1281,3 @@ msgstr "หมายเหตุ"
 #, python-format
 msgid "Partner is required on each line."
 msgstr "ต้องระบุผู้รับเงินในทุกบรรทัด"
-
-#. module: disbursement
-msgid "Disbursement Assignment Rule"
-msgstr "กฎการมอบหมายใบขอเบิก"
-
-#. module: disbursement
-msgid "Assign Disbursement Verification Officer"
-msgstr "มอบหมายเจ้าหน้าที่หมวดตรวจใบขอเบิก"
-
-#. module: disbursement
-#: model:ir.actions.act_window,name:disbursement.action_disbursement_assignment_rule
-#: model:ir.ui.menu,name:disbursement.menu_disbursement_assignment_rule
-msgid "Assignment Rules"
-msgstr "กฎการมอบหมาย"
-
-#. module: disbursement
-#: model:ir.model.fields,field_description:disbursement.field_disbursement_assignment_rule__user_id
-msgid "Assigned Officer"
-msgstr "เจ้าหน้าที่ผู้รับผิดชอบ"
-
-#. module: disbursement
-msgid "Assign to me"
-msgstr "รับงานนี้"
-
-#. module: disbursement
-msgid "Assign…"
-msgstr "มอบหมาย…"
-
-#. module: disbursement
-msgid "Unassign"
-msgstr "ถอนผู้รับผิดชอบ"
-
-#. module: disbursement
-msgid "Assign"
-msgstr "มอบหมาย"
-
-#. module: disbursement
-msgid "Assign Officer"
-msgstr "มอบหมายเจ้าหน้าที่"
-
-#. module: disbursement
-msgid "My Verifications"
-msgstr "งานตรวจของฉัน"
-
-#. module: disbursement
-msgid "Unassigned"
-msgstr "ยังไม่มีผู้รับผิดชอบ"
-
-#. module: disbursement
-msgid "Assigned as responsible verification officer"
-msgstr "ได้รับมอบหมายเป็นเจ้าหน้าที่หมวดตรวจผู้รับผิดชอบ"
-
-#. module: disbursement
-msgid "Apply Rules to Pending Requests"
-msgstr "นำกฎไปใช้กับใบขอเบิกที่รอตรวจ"
-
-#. module: disbursement
-msgid "Assignment Rules Applied"
-msgstr "นำกฎการมอบหมายไปใช้แล้ว"
-
-#. module: disbursement
-#, python-format
-msgid "%(assigned)s of %(total)s pending request(s) were assigned."
-msgstr "มอบหมายแล้ว %(assigned)s จาก %(total)s ใบที่รอตรวจ"
-
-#. module: disbursement
-#, python-format
-msgid "This document is already assigned to %s."
-msgstr "เอกสารนี้มีผู้รับผิดชอบแล้วคือ %s"
-
-#. module: disbursement
-msgid "Only a manager can unassign the officer."
-msgstr "เฉพาะผู้จัดการเท่านั้นที่ถอนผู้รับผิดชอบได้"
-
-#. module: disbursement
-msgid "Only a manager can assign another officer."
-msgstr "เฉพาะผู้จัดการเท่านั้นที่มอบหมายผู้รับผิดชอบคนอื่นได้"
-
-#. module: disbursement
-msgid "Disbursement Verification"
-msgstr "หมวดตรวจ"
-
-#. module: disbursement
-#: model:ir.ui.menu,name:disbursement.menu_disbursement_verification
-msgid "Verification"
-msgstr "หมวดตรวจ"
-
-#. module: disbursement
-#: model:ir.actions.act_window,name:disbursement.action_disbursement_my_verifications
-#: model:ir.ui.menu,name:disbursement.menu_disbursement_my_verifications
-msgid "My Work"
-msgstr "งานของฉัน"
-
-#. module: disbursement
-#: model:ir.actions.act_window,name:disbursement.action_disbursement_all
-#: model:ir.ui.menu,name:disbursement.menu_disbursement_all_requests
-msgid "All Disbursement Requests"
-msgstr "ใบขอเบิกทั้งหมด"
-
-#. module: disbursement
-msgid "No disbursement request is assigned to you."
-msgstr "ยังไม่มีใบขอเบิกที่มอบหมายให้คุณ"
-
-#. module: disbursement
-msgid "Return"
-msgstr "ตีกลับ"
-
-#. module: disbursement
-msgid "Return for Correction"
-msgstr "ตีกลับเพื่อแก้ไข"
-
-#. module: disbursement
-msgid "Returned for Correction"
-msgstr "ถูกตีกลับเพื่อแก้ไข"
-
-#. module: disbursement
-msgid "Reason"
-msgstr "เหตุผล"
-
-#. module: disbursement
-msgid "Reason for returning this request…"
-msgstr "ระบุเหตุผลที่ตีกลับใบขอเบิกนี้…"
-
-#. module: disbursement
-msgid "Resubmit to Verification"
-msgstr "ส่งกลับหมวดตรวจ"
-
-#. module: disbursement
-#, python-format
-msgid "Returned for correction: %s"
-msgstr "ตีกลับเพื่อแก้ไข: %s"
-
-#. module: disbursement
-msgid "Returned for correction"
-msgstr "ตีกลับเพื่อแก้ไข"
-
-#. module: disbursement
-msgid "Corrected and returned to verification."
-msgstr "แก้ไขแล้ว ส่งกลับหมวดตรวจ"
-
-#. module: disbursement
-#, python-format
-msgid "Only a request under verification (signed) can be returned."
-msgstr "ตีกลับได้เฉพาะใบที่อยู่ระหว่างหมวดตรวจ (สถานะ signed) เท่านั้น"
-
-#. module: disbursement
-#, python-format
-msgid "Only a returned request in draft can be resubmitted."
-msgstr "ส่งกลับได้เฉพาะใบที่ถูกตีกลับและอยู่สถานะ draft เท่านั้น"
-
-#. module: disbursement
-#, python-format
-msgid "Cannot resubmit a disbursement request with no lines."
-msgstr "ส่งกลับไม่ได้ เพราะใบขอเบิกไม่มีรายการ"
