+++ disbursement/security/ir.model.access.csv
@@ -9,7 +9,3 @@ access_disbursement_request_line_head,disbursement.request.line.head,model_disbu
 access_disbursement_request_line_officer,disbursement.request.line.officer,model_disbursement_request_line,group_disbursement_officer,1,1,1,0
 access_disbursement_request_line_manager,disbursement.request.line.manager,model_disbursement_request_line,group_disbursement_manager,1,1,1,1
 access_disbursement_exception_confirm,disbursement.exception.confirm,model_disbursement_exception_confirm,,1,1,1,1
-access_disbursement_assignment_rule_user,disbursement.assignment.rule user,model_disbursement_assignment_rule,group_disbursement_user,1,0,0,0
-access_disbursement_assignment_rule_manager,disbursement.assignment.rule manager,model_disbursement_assignment_rule,group_disbursement_manager,1,1,1,1
-access_disbursement_assign_officer_wizard,disbursement.assign.officer.wizard,model_disbursement_assign_officer_wizard,group_disbursement_officer,1,1,1,1
-access_disbursement_return_request_wizard,disbursement.return.request.wizard,model_disbursement_return_request_wizard,group_disbursement_officer,1,1,1,1
