+++ disbursement/wizards/__init__.py
@@ -1,5 +1,3 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
 from . import disbursement_exception_confirm
-from . import assign_officer_wizard
-from . import return_request_wizard
