+++ disbursement_assignment_kmitl/__init__.py
@@ -1,2 +1,4 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
+from . import models
+from . import wizards
