+++ disbursement_assignment_kmitl/i18n/th.po
@@ -1,0 +1,203 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+# 	* disbursement_assignment_kmitl
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 16.0\n"
+"Report-Msgid-Bugs-To: \n"
+"POT-Creation-Date: 2026-03-19 10:37+0000\n"
+"PO-Revision-Date: 2026-03-19 17:56+0700\n"
+"Last-Translator: \n"
+"Language-Team: \n"
+"Language: th\n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: 8bit\n"
+"Plural-Forms: nplurals=1; plural=0;\n"
+"X-Generator: Poedit 3.8\n"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__active
+msgid "Active"
+msgstr "ใช้งาน"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__activity_analytic_id
+msgid "Activity"
+msgstr "กิจกรรม"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__department_analytic_id
+msgid "Department"
+msgstr "ส่วนงาน"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__fund_analytic_id
+msgid "Fund"
+msgstr "กองทุน"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__partner_type_id
+msgid "Partner Type"
+msgstr "จ่ายตาม"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__source_analytic_id
+msgid "Source"
+msgstr "แหล่งเงิน"
+
+#. module: disbursement_assignment_kmitl
+msgid "Disbursement Assignment Rule"
+msgstr "กฎการมอบหมายใบขอเบิก"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assign Disbursement Verification Officer"
+msgstr "มอบหมายเจ้าหน้าที่หมวดตรวจใบขอเบิก"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.actions.act_window,name:disbursement_assignment_kmitl.action_disbursement_assignment_rule
+#: model:ir.ui.menu,name:disbursement_assignment_kmitl.menu_disbursement_assignment_rule
+msgid "Assignment Rules"
+msgstr "กฎการมอบหมาย"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.model.fields,field_description:disbursement_assignment_kmitl.field_disbursement_assignment_rule__user_id
+msgid "Assigned Officer"
+msgstr "เจ้าหน้าที่ผู้รับผิดชอบ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assign to me"
+msgstr "รับงานนี้"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assign…"
+msgstr "มอบหมาย…"
+
+#. module: disbursement_assignment_kmitl
+msgid "Unassign"
+msgstr "ถอนผู้รับผิดชอบ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assign"
+msgstr "มอบหมาย"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assign Officer"
+msgstr "มอบหมายเจ้าหน้าที่"
+
+#. module: disbursement_assignment_kmitl
+msgid "My Verifications"
+msgstr "งานตรวจของฉัน"
+
+#. module: disbursement_assignment_kmitl
+msgid "Unassigned"
+msgstr "ยังไม่มีผู้รับผิดชอบ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assigned as responsible verification officer"
+msgstr "ได้รับมอบหมายเป็นเจ้าหน้าที่หมวดตรวจผู้รับผิดชอบ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Apply Rules to Pending Requests"
+msgstr "นำกฎไปใช้กับใบขอเบิกที่รอตรวจ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Assignment Rules Applied"
+msgstr "นำกฎการมอบหมายไปใช้แล้ว"
+
+#. module: disbursement_assignment_kmitl
+#, python-format
+msgid "%(assigned)s of %(total)s pending request(s) were assigned."
+msgstr "มอบหมายแล้ว %(assigned)s จาก %(total)s ใบที่รอตรวจ"
+
+#. module: disbursement_assignment_kmitl
+#, python-format
+msgid "This document is already assigned to %s."
+msgstr "เอกสารนี้มีผู้รับผิดชอบแล้วคือ %s"
+
+#. module: disbursement_assignment_kmitl
+msgid "Only a manager can unassign the officer."
+msgstr "เฉพาะผู้จัดการเท่านั้นที่ถอนผู้รับผิดชอบได้"
+
+#. module: disbursement_assignment_kmitl
+msgid "Only a manager can assign another officer."
+msgstr "เฉพาะผู้จัดการเท่านั้นที่มอบหมายผู้รับผิดชอบคนอื่นได้"
+
+#. module: disbursement_assignment_kmitl
+msgid "Disbursement Verification"
+msgstr "หมวดตรวจ"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.ui.menu,name:disbursement_assignment_kmitl.menu_disbursement_verification
+msgid "Verification"
+msgstr "หมวดตรวจ"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.actions.act_window,name:disbursement_assignment_kmitl.action_disbursement_my_verifications
+#: model:ir.ui.menu,name:disbursement_assignment_kmitl.menu_disbursement_my_verifications
+msgid "My Work"
+msgstr "งานของฉัน"
+
+#. module: disbursement_assignment_kmitl
+#: model:ir.actions.act_window,name:disbursement_assignment_kmitl.action_disbursement_all
+#: model:ir.ui.menu,name:disbursement_assignment_kmitl.menu_disbursement_all_requests
+msgid "All Disbursement Requests"
+msgstr "ใบขอเบิกทั้งหมด"
+
+#. module: disbursement_assignment_kmitl
+msgid "No disbursement request is assigned to you."
+msgstr "ยังไม่มีใบขอเบิกที่มอบหมายให้คุณ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Return"
+msgstr "ตีกลับ"
+
+#. module: disbursement_assignment_kmitl
+msgid "Return for Correction"
+msgstr "ตีกลับเพื่อแก้ไข"
+
+#. module: disbursement_assignment_kmitl
+msgid "Returned for Correction"
+msgstr "ถูกตีกลับเพื่อแก้ไข"
+
+#. module: disbursement_assignment_kmitl
+msgid "Reason"
+msgstr "เหตุผล"
+
+#. module: disbursement_assignment_kmitl
+msgid "Reason for returning this request…"
+msgstr "ระบุเหตุผลที่ตีกลับใบขอเบิกนี้…"
+
+#. module: disbursement_assignment_kmitl
+msgid "Resubmit to Verification"
+msgstr "ส่งกลับหมวดตรวจ"
+
+#. module: disbursement_assignment_kmitl
+#, python-format
+msgid "Returned for correction: %s"
+msgstr "ตีกลับเพื่อแก้ไข: %s"
+
+#. module: disbursement_assignment_kmitl
+msgid "Returned for correction"
+msgstr "ตีกลับเพื่อแก้ไข"
+
+#. module: disbursement_assignment_kmitl
+msgid "Corrected and returned to verification."
+msgstr "แก้ไขแล้ว ส่งกลับหมวดตรวจ"
+
+#. module: disbursement_assignment_kmitl
+#, python-format
+msgid "Only a request under verification (signed) can be returned."
+msgstr "ตีกลับได้เฉพาะใบที่อยู่ระหว่างหมวดตรวจ (สถานะ signed) เท่านั้น"
+
+#. module: disbursement_assignment_kmitl
+#, python-format
+msgid "Only a returned request in draft can be resubmitted."
+msgstr "ส่งกลับได้เฉพาะใบที่ถูกตีกลับและอยู่สถานะ draft เท่านั้น"
+
+#. module: disbursement_assignment_kmitl
+#, python-format
+msgid "Cannot resubmit a disbursement request with no lines."
+msgstr "ส่งกลับไม่ได้ เพราะใบขอเบิกไม่มีรายการ"
+
