+++ disbursement_assignment_kmitl/models/__init__.py
@@ -1,2 +1,4 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
+from . import assignment_rule
+from . import disbursement_request
