+++ disbursement_assignment_kmitl/tests/__init__.py
@@ -1,2 +1,3 @@
 # License LGPL-3.0 or later (https://www.gnu.org/licenses/lgpl).
 
+from . import test_assignment_rule
